[Multiplayer] Proxy Colors on just spawned entities now work.
Networking over the entities index value instead of itself so it can be checked multiple times over time to see if it becomes valid. Makes duplications on servers with proxy colors work. Applying proxy colors on an entity that just spawned now works too! (included a fallback for a 3 second check delay just in case)

Code may be unneeded or redundent but I dont care as it now works and should work in most conditions.